Respond correctly to failed DeliverSM
The SMPP specification specifically lists that a PDU body with a NULL field message_id is required in the deliver_sm_resp message. Using the sendNegativeResponse() will result in a body-less PDU response that some SMPP servers legitimately refuse with an error about invalid command length. Now we correctly send the deliver_sm_resp in case of a ProcessRequestException. For PDUStringException we now send a generick_nack instead of a deliver_sm_resp, just like in other cases were such an exception is thrown.